bring models into news app
