--- base/__init__.py
@@ -1,2 +1,0 @@
-#!/usr/bin/env python3
-"""Provides common functionality to share amongst apps"""
