--- comments/models/__init__.py
@@ -1,2 +1,0 @@
-#!/usr/bin/env python3
-"""Provides models for the comments app"""
